test(charts): relocate unit tests into __tests__ so CI runs them; drop dead import
