+++ packages/web/src/components/charts/__tests__/trend-utils.test.ts
@@ -8,7 +8,7 @@ import {
   buildTrendChartData,
   personalBestIndex,
   radarDataFromPercentiles,
-} from './trend-utils';
+} from '../trend-utils';
 import type { BenchmarkOverlay, MetricTrend } from '@shared/report-trends-types';
 
 describe('trend-utils', () => {
